moves files to future use folder, reworked Creature and character into a IUnit instead
