+++ Assets/Scripts/GameData/Equipment/IArnor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3ef74f8feb66d84fa993e72ed905709
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
