+++ Assets/Scripts/GameData/Equipment/ISpellBook.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6f8561744336f04aaba20c96cc3cf76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
