+++ Assets/Scripts/GameData/Equipment/IWeapon.cs
@@ -1,5 +1,4 @@
 ﻿using SwordAndBored.GameData.Abilities;
-using SwordAndBored.GameData.Modifiers;
 
 namespace SwordAndBored.GameData.Equipment
 {
