+++ Assets/Scripts/GameData/Equipment/IWeapon.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f1327e84e6e081d4b9e368d5a730197a
+guid: 06fe244221e138046a1696351638436a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
