+++ Assets/Scripts/GameData/Modifiers/IModifierDefense.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b1026883e9d5b541bcd5cb6b162aa6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
