+++ Assets/Scripts/GameData/Modifiers/IModifierStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7e84e61c7aee0b149863ebfb7db1e0f8
+guid: 47123051835124448b3e21d4deb9e8f8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
