+++ Assets/Scripts/GameData/Unused For Now/GenericCharacter.cs
@@ -1,4 +1,5 @@
-﻿using SwordAndBored.GameData.Abilities;
+﻿/*
+using SwordAndBored.GameData.Abilities;
 using SwordAndBored.GameData.Abilities.Skills;
 using SwordAndBored.GameData.Roles;
 using System;
@@ -51,3 +52,4 @@ namespace SwordAndBored.GameData.Creatures
         }
     }
 }
+*/
