+++ Assets/Scripts/GameData/Unused For Now/GenericCreature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 539af6bf9dcf70a459c72b0c526b8063
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
