+++ Assets/Scripts/GameData/Unused For Now/IHero.cs
